Make signing groups opt in instead of opt out
opt out is bad for many reasons, syncs fail unless you opt out if you don't have the option, when you do enable it and you have an in progress sync erroring it starts erroring on unknow resource. This requires support dashboard to fix. It also fails the sync because it is a bad sync if you disable it once you had it.